requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis & Celery
redis>=5.0.0
celery[redis]>=5.3.0

# Security
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0
